Move BusController to its own package
